requirements.txt
fastapi==0.81.0
uvicorn==0.18.3
Levenshtein==0.20.2
psycopg2-binary==2.9.3
python-multipart==0.0.5